server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.129.0",
    "nats-py>=2.13.1",
    "passlib>=1.7.4",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.46",
    "uuid6>=2025.0.1",
    "uvicorn>=0.41.0",
    "uvloop>=0.22.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
